Rename jsxSpreadAttribute to objectAssign
This is so that I can use the same runtime funtion for:

https://github.com/sebmarkbage/ecmascript-rest-spread/blob/master/Spec.md
